+++ src/components/basic_table/basic_table.test.js
@@ -99,6 +99,7 @@ describe('EuiBasicTable', () => {
         { id: '2', name: 'name2' },
         { id: '3', name: 'name3' }
       ],
+      itemId: 'id',
       columns: [
         {
           field: 'name',